+++ CHANGELOG.rst
@@ -3,7 +3,9 @@
 * `#566 <https://github.com/xitrum-framework/xitrum/issues/566>`_
   Update Scaposer from 1.6 to 1.7
 * `#572 <https://github.com/xitrum-framework/xitrum/issues/572>`_
-  JavaScript error at metrics detail page of actions
+  [Metrics] Fix JavaScript error at metrics detail page of actions
+* `#574 <https://github.com/xitrum-framework/xitrum/issues/574>`_
+  [Metrics] Add single executionTime histogram for all actions
 
 3.25.0: